refactor(queuev2): extract cachedQueueReaderBase
Move the state and behaviour shared by cached queue readers into a new
cachedQueueReaderBase struct: the cached window, coverage checks, rangeID
fencing, the read path (including shadow-mode comparison), eviction bound
bookkeeping, the daemon lifecycle status, the common cachedQueueReaderOptions,
and the single CachedQueueReader interface.

The scheduled (timer) reader now embeds *cachedQueueReaderBase and keeps only
its prefetch-specific state and behaviour. Field and method promotion leaves
existing call sites and the timer test unchanged.

Keep the base generic so a second reader can embed it later:
- The prefetch-only fields (prefetchTargetUpper, pendingInjectBuffer) stay on
  the scheduled reader, not the base.
- The base carries a default LookAHead that delegates to the DB reader; the
  scheduled reader overrides it with its cache-aware version.
- The base exposes two nil-safe hooks run under mu: onUpperBoundUpdatedLockedFn
  (scheduled wires it to notifyPrefetch) and onClearLockedFn (scheduled wires it
  to a new resetPrefetchState so the shared clearLocked still discards
  prefetch-only state).
- The shared cachedQueueReaderState snapshot drops the prefetch-only target
  bound; the scheduled reader instead logs the prefetch target on its
  buffering-task debug line.

Behavior-preserving structural move.
